Replaced Mailchimp component with Recapcha component
